v1 actDatosPer y Avance-1 actDatosEsc

+++ PHP/actDatosPer.php
@@ -1,4 +1,9 @@
 <?php
+if(!isset($_FILES['foto']['name'])){
+	session_start();
+	$_SESSION['error']=1;
+	header("location:../HTML/actDatosPer.html");
+}else{
     $actAlumno = true;
 	$uploadedfileload=true;
 	//Aquí se obtiene la extensión del archivo
@@ -6,13 +11,12 @@
 	$ext = pathinfo($path, PATHINFO_EXTENSION); 
 	//Fin
 	//Aquí se valida si se selecciono una foto
-	if ($_FILES['foto']["error"] > 0){
-		//echo "<h3 style='color:red;'>Error: " . $_FILES['foto']['error']."<h3>";
+	if ($_FILES['foto']['error'] > 0){
+		//echo "<h3 style='color:red;'>Error: " . $_FILES['foto']['error']."</h3>";
 		$uploadedfileload=false;
 	}
 	if($ext==""){
 		$uploadedfileload=false;
-		echo "<h3 style='color: red;'>No agrego foto</h3>";
 	//Fin
 	}else{
 		//Aquí se valida que la imagen/foto tenga las caracteristicas correspondientes (fomato: jpg y peso<=200KB)
@@ -28,7 +32,7 @@
 		if(!$uploadedfileload){
 			$actAlumno = false;
 			echo "<h3 style='color: red;'>".$msg."</h3>";
-			echo "<a href='../HTML/registroAlumno.html'>Volver a los datos</a>";
+			echo "<a href='../HTML/actDatosPer.html'>Regresar a los Datos</a>";
 		}
         //Fin
     }
@@ -55,53 +59,70 @@
 				}
 			}
 		}else{
+			$actAlumno = false;
 			echo "<h3 style='color: red;'>No se pudieron actualizar los datos. Inténtelo de nuevo.</h3>";
-			echo "<a href='../HTML/principalUsuario.html'>Volver a principal</a>";
 		}
 		if($actAlumno){
-			echo "Se pueden actualizar los datos";
 			$consultaAlumno=mysqli_query($conexion,"SELECT apepat,apemat,nombre,foto FROM datospersonales WHERE clave='$idAlumno'");
 			if(mysqli_num_rows($consultaAlumno)){
 				$info=mysqli_fetch_array($consultaAlumno);
 				$nombreBD=$info['apepat'].$info['apemat'].$info['nombre'];
 				$nombreAct=$apepat.$apemat.$nombre;
-				echo "<hr>";
-				echo $nombreBD."<br>";
-				echo $nombreAct."<br>";
 				(strcmp($nombreBD,$nombreAct)==0) ? $nomIgual=true : $nomIgual=false;
-				if($uploadedfileload){
-					//Aquí se pone la extensión en el nombre de la Foto
-					$file_name=$apepat." ".$apemat." ".$nombre.".".$ext;	
-					//Fin
-					$path="./../Fotos/".$file_name;
-					echo "Ruta del archivo:".$path;
-					if(move_uploaded_file ($_FILES['foto']['tmp_name'], $path)){
-						$subirFoto=mysqli_query($conexion,"UPDATE datospersonales SET foto='$path' WHERE clave='$idAlumno'");
-						if($subirFoto){
-							if($info['foto']!=null && !$nomIgual){
-								$foto_anterior=$info['apepat']." ".$info['apemat']." ".$info['nombre'].".".$ext;
-									if(file_exists("./../Fotos/".$foto_anterior)){
-										if(unlink("./../Fotos/".$foto_anterior)){
-											echo "<h3>La Foto Anterior se Ha Eliminado</h3>";
-										}else{
-											echo "<h3>Favor de Eliminar la Foto Anterior</h3>";
-										}
+				if($_FILES['foto']['error']<=0){
+						if($uploadedfileload){
+							//Aquí se pone la extensión en el nombre de la Foto
+							$file_name=$apepat." ".$apemat." ".$nombre.".".$ext;	
+							//Fin
+							$path="./../Fotos/".$file_name;
+							if(move_uploaded_file($_FILES['foto']['tmp_name'], $path)){
+								$subirFoto=mysqli_query($conexion,"UPDATE datospersonales SET foto='$path' WHERE clave='$idAlumno'");
+								if($subirFoto){
+									if($info['foto']!=null && !$nomIgual){
+										$foto_anterior=$info['apepat']." ".$info['apemat']." ".$info['nombre'].".".$ext;
+											if(file_exists("./../Fotos/".$foto_anterior)){
+												if(unlink("./../Fotos/".$foto_anterior)){
+													echo "<h3>La Foto Anterior se Ha Eliminado</h3>";
+												}else{
+													echo "<h3>Favor de Eliminar la Foto Anterior</h3>";
+												}
+											}else{
+												echo "<h3 style='color: red;'>No se encontró la Foto</h3>";
+											}
 									}
+									echo "<h3 style='color: green;'>La Foto se Ha Subido Satisfactoriamente</h3>";
+								}else{
+									echo "<h3 style='color: red;'>No se pudo Subir la Foto. Favor de Intentarlo de Nuevo</h3>";
+								}
+							}else{
+								echo "<h3 style='color: red;'>No se pudo Subir la Foto. Favor de Intentarlo de Nuevo.</h3>";
 							}
-							echo "<h3 style='color: green;'>La Foto se Ha Subido Satisfactoriamente</h3>";
+						}
+				}
+				if(!$uploadedfileload && !$nomIgual){
+					$file_name=$apepat." ".$apemat." ".$nombre.".jpg";
+					$path="./../Fotos/".$file_name;
+					if(file_exists($info['foto'])){
+						if(rename($info['foto'],"./../Fotos/".$file_name)){
+							$camNomFoto = mysqli_query($conexion,"UPDATE datospersonales SET foto='$path' WHERE clave='$idAlumno'");
+							if($camNomFoto){
+								echo "<h3 style='color: green;'>La Foto fue Renombrada con Éxito</h3>";
+							}else{
+								echo "<h3 style='color: red;'>La Foto no Cambio su Nombre</h3>";
+							}
 						}else{
-							echo "<h3 style='color: red;'>No se pudo Subir la Foto</h3>";
-							echo "<h3>Favor de Intentarlo de Nuevo</h3>";
+							echo "<h3 style='color: red;'>La Foto no Cambio su Nombre</h3>";
 						}
 					}else{
-						echo "<h3 style='color: red;'>No se pudo Subir la Foto</h3>";
-						echo "<h3>Favor de Intentarlo de Nuevo</h3>";
+						echo "<h3 style='color: red;'>No se encontró la Foto</h3>";
 					}
 				}
 			}else{
-				echo "<h3 style='color: red;'>No se pudo Subir la Foto. Inténtelo de Nuevo</h3><br>
-						<a href='../HTML/principalUsuario.html>Volver a principal</a>'";
+				if($_FILES['foto']['error']<=0){
+					echo "<h3 style='color: red;'>No se pudo Actualizar la Foto. Inténtelo de Nuevo</h3>";
+				}
 			}
+
 			$genero = $_POST['genero'];
 			$fnac = $_POST['fnac'];
 			$direccion = trim($_POST['direccion']);
@@ -115,12 +136,15 @@
 			WHERE clave='$idAlumno'");
 			if($actualizarAlumno){
 				echo "<h3 style='color: green;'>DATOS ACTUALIZADOS CORRECTAMENTE</h3>";
+				echo "<a href='../HTML/actDatosPer.html'>Ver los Datos del Alumno</a>";
 			}else{
-				echo "<h3 style='color: red;'>No se pudieron actualizar los datos</h3>";
+				echo "<h3 style='color: red;'>No se pudieron actualizar los datos. Inténtelo de Nuevo</h3>";
+				echo "<a href='../HTML/actDatosPer.html'>Volver a los Datos del Alumno</a>";
 			}
 		}else{
 			echo "<h3 style='color: red;'>Ya existe un Alumno con esa CURP y/o Nombre</h3>";
-			echo "<a href='../HTML/principalUsuario.html'>Volver a principal</a>";
+			echo "<a href='../HTML/actDatosPer.html'>Volver a los Datos del Alumno</a>";
 		}
-    }
+	}
+}
 ?>
